HRMS-BOBO-BE/requirements.txt
django
djangorestframework
pyyaml
django-cors-headers
pillow
obfuskey
django-filter
psycopg2
django-crum
django-mirage-field
openpyxl
django_import_export
daphne
openai
urllib3<2.0
Twisted[tls,http2]
requests
django-ckeditor
num2words
pandas
pymysql
pyotp
qrcode[pil]
pyzk